+++ NEWS.md
@@ -1,4 +1,23 @@
-# SEAPSUTWorkflow 0.2.3 (2021-08-20)
+---
+title: "Release notes for `SEAPSUTWorkflow`"
+output: html_document
+---
+
+
+# SEAPSUTWorkflow 0.2.4 (2021-10-15) 
+
+* Fixed a bug where the primary-to-final exergy-to-energy ratio
+  (phi) didn't get row and column types.
+* New targets:
+    * `phi_constants_path` gives path to a phi constants file.
+    * `Phi_constants` is a table of reasonable constant values of phi (the energy-to-exergy) ratio.
+* Change name of constant: `row_code` --> `region_code`.
+* New tests for new features.
+  * Now up to 306 tests, all passing.
+  * Test coverage remains at 100%.
+
+
+# SEAPSUTWorkflow 0.2.3 (2021-08-20) [![DOI](https://zenodo.org/badge/265032888.svg)](https://zenodo.org/badge/latestdoi/265032888)
 
 * Now using "WLD" instead of "World" everywhere.
 * New targets: `country_concordance_path1 and `CountryConcordanceTable`.